\name{getcdfupdates-methods}
\docType{methods}
\alias{getcdfupdates}
\alias{getcdfupdates-methods}
\alias{getcdfupdates,SPCCUSUMNonpar-method}
\alias{getcdfupdates,SPCCUSUMNormal-method}
\alias{getcdfupdates,SPCShewNonparCenterScale-method}
\alias{getcdfupdates,SPCShewNormalCenterScale-method}
\title{Methods for Function \code{getcdfupdates} in Package \pkg{spcadjust}}
\description{
Methods for function \code{getcdfupdates} in package \pkg{spcadjust} 
}
\section{Methods}{
\describe{

\item{\code{signature(chart = "SPCCUSUMNonpar")}}{
}

\item{\code{signature(chart = "SPCCUSUMNormal")}}{
}

\item{\code{signature(chart = "SPCShewNonparCenterScale")}}{
}

\item{\code{signature(chart = "SPCShewNormalCenterScale")}}{
}
}}
\keyword{methods}
