\name{Pofdata-methods}
\docType{methods}
\alias{Pofdata}
\alias{Pofdata-methods}
\alias{Pofdata,SPCCUSUMNonpar-method}
\alias{Pofdata,SPCCUSUMNormal-method}
\alias{Pofdata,SPCShewNonparCenterScale-method}
\alias{Pofdata,SPCShewNormalCenterScale-method}
\title{Methods for Function \code{Pofdata} in Package \pkg{spcadjust} }
\description{
Methods for function \code{Pofdata} in package \pkg{spcadjust} 
}
\section{Methods}{
\describe{

\item{\code{signature(chart = "SPCCUSUMNonpar")}}{
}

\item{\code{signature(chart = "SPCCUSUMNormal")}}{
}

\item{\code{signature(chart = "SPCShewNonparCenterScale")}}{
}

\item{\code{signature(chart = "SPCShewNormalCenterScale")}}{
}
}}
\keyword{methods}
