\name{resample-methods}
\docType{methods}
\alias{resample}
\alias{resample-methods}
\alias{resample,SPCCUSUMNonpar-method}
\alias{resample,SPCCUSUMNormal-method}
\alias{resample,SPCShewNonparCenterScale-method}
\alias{resample,SPCShewNormalCenterScale-method}
\title{Methods for Function \code{resample} in Package \pkg{spcadjust}}
\description{
Methods for function \code{resample} in package \pkg{spcadjust}
}
\section{Methods}{
\describe{

\item{\code{signature(chart = "SPCCUSUMNonpar")}}{
}

\item{\code{signature(chart = "SPCCUSUMNormal")}}{
}

\item{\code{signature(chart = "SPCShewNonparCenterScale")}}{
}

\item{\code{signature(chart = "SPCShewNormalCenterScale")}}{
}
}}
\keyword{methods}
