\name{runchart-methods}
\docType{methods}
\alias{runchart}
\alias{runchart-methods}
\alias{runchart,SPCCUSUM-method}
\alias{runchart,SPCCUSUM-method,ANY,ANY}
\alias{runchart,SPCShew-method}
\title{Methods for Function \code{runchart} in Package \pkg{spcadjust}.}
\description{
Runs the chart.
}
\section{Methods}{
\describe{

\item{\code{signature(chart = "SPCCUSUM")}}{
}

\item{\code{signature(chart = "SPCShew")}}{
}
}}
\keyword{methods}
