% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{class}
\name{SPCchart-class}
\alias{SPCchart-class}
\title{Virtual Base Class for Control Charts}
\description{
Virtual S4 base class for Control Charts.
}
\section{Slots}{

\describe{
\item{\code{model}}{The data model to be used in the chart. Must be of type
\code{\link{SPCDataModel}}.}
}}

