\name{SPCoutput-methods}
\docType{methods}
\alias{SPCoutput}
\alias{SPCoutput-methods}
\alias{SPCoutput,SPCARL-method}
\alias{SPCoutput,SPCcalARL-method}
\alias{SPCoutput,SPCcalhitprob-method}
\alias{SPCoutput,SPChitprob-method}
\title{Methods for Function \code{SPCoutput} in Package \pkg{spcadjust}}
\description{
Methods for function \code{SPCoutput} in package \pkg{spcadjust}
}
\section{Methods}{
\describe{

\item{\code{signature(property = "SPCARL")}}{
}

\item{\code{signature(property = "SPCcalARL")}}{
}

\item{\code{signature(property = "SPCcalhitprob")}}{
}

\item{\code{signature(property = "SPChitprob")}}{
}
}}
\keyword{methods}
