% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{class}
\name{SPCpropertyres-class}
\alias{SPCpropertyres-class}
\alias{show,SPCpropertyres-method}
\title{Results of SPCproperty.}
\usage{
\S4method{show}{SPCpropertyres}(object)
}
\arguments{
\item{object}{the result to be shown.}
}
\description{
Results of SPCproperty.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Prints the object nicely.
}}
\section{Slots}{

\describe{
\item{\code{nrep}}{number of repetitions used in the simulation.}

\item{\code{chart}}{the chart used.}

\item{\code{property}}{the property of interest, \code{ARL}, \code{calARL}, \code{calhitprob} or \code{hitprob}.}

\item{\code{covprob}}{the probability of the guaranteed conditional performance.}

\item{\code{res}}{the guaranteed conditional performance.}

\item{\code{raw}}{the unadjusted result.}

\item{\code{params}}{additional parameters used for computing this property.}

\item{\code{restext}}{a readable version of the result.}
}}

