\name{SPCq-methods}
\docType{methods}
\alias{SPCq}
\alias{SPCq-methods}
\alias{SPCq,ARLCUSUM-method}
\alias{SPCq,ARLShew-method}
\alias{SPCq,calARLCUSUM-method}
\alias{SPCq,calARLShew-method}
\alias{SPCq,calhitprobCUSUM-method}
\alias{SPCq,calhitprobShew-method}
\alias{SPCq,hitprobCUSUM-method}
\alias{SPCq,hitprobShew-method}
\title{Methods for Function \code{SPCq} in Package \pkg{spcadjust}}
\description{
Methods for function \code{SPCq} in package \pkg{spcadjust}
}
\section{Methods}{
\describe{

\item{\code{signature(property = "ARLCUSUM")}}{
}

\item{\code{signature(property = "ARLShew")}}{
}

\item{\code{signature(property = "calARLCUSUM")}}{
}

\item{\code{signature(property = "calARLShew")}}{
}

\item{\code{signature(property = "calhitprobCUSUM")}}{
}

\item{\code{signature(property = "calhitprobShew")}}{
}

\item{\code{signature(property = "hitprobCUSUM")}}{
}

\item{\code{signature(property = "hitprobShew")}}{
}
}}
\keyword{methods}
