\name{updates-methods}
\docType{methods}
\alias{updates}
\alias{updates-methods}
\alias{updates,SPCCUSUMNonparCenterScale-method}
\alias{updates,SPCCUSUMNormal-method}
\alias{updates,SPCShewNonparCenterScale-method}
\alias{updates,SPCShewNormalCenterScale-method}
\title{Methods for Function \code{updates} in Package \pkg{spcadjust}}
\description{
 Methods for function \code{updates} in package \pkg{spcadjust} 
}
\section{Methods}{
\describe{

\item{\code{signature(chart = "SPCCUSUMNonparCenterScale")}}{
}

\item{\code{signature(chart = "SPCCUSUMNormal")}}{
}

\item{\code{signature(chart = "SPCShewNonparCenterScale")}}{
}

\item{\code{signature(chart = "SPCShewNormalCenterScale")}}{
}
}}
\keyword{methods}
