% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R, R/shewhart.R
\docType{methods}
\name{updates}
\alias{updates}
\alias{updates,SPCShew-method}
\alias{updates,SPCchart-method}
\title{Updates of a Control Chart}
\usage{
updates(chart, xi, data)

\S4method{updates}{SPCchart}(chart, xi, data)

\S4method{updates}{SPCShew}(chart, xi, data)
}
\arguments{
\item{chart}{the control chart.}

\item{xi}{the parameters used for running the chart.}

\item{data}{the observed data.}
}
\value{
A vector of the same length as data.
}
\description{
Computes updates of a control chart using the given parameters and the given data.
}
\section{Methods (by class)}{
\itemize{
\item \code{SPCchart}: Standard method which simply first applies
getresiduals from the data model.

\item \code{SPCShew}: Computes the updates taking into account if the
chart is one-sided or two-sided.
}}

