\name{xiofdata-methods}
\docType{methods}
\alias{xiofdata}
\alias{xiofdata-methods}
\alias{xiofdata,SPCchart-method}
\title{Methods for Function \code{xiofdata} in Package \pkg{spcadjust}}
\description{
Methods for function \code{xiofdata} in package \pkg{spcadjust} 
}
\section{Methods}{
\describe{

\item{\code{signature(chart = "SPCchart")}}{
}
}}
\keyword{methods}
