% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{methods}
\name{xiofdata}
\alias{xiofdata}
\alias{xiofdata,SPCchart-method}
\title{Estimate Chart Parameters}
\usage{
xiofdata(chart, data)

\S4method{xiofdata}{SPCchart}(chart, data)
}
\arguments{
\item{chart}{the control chart to be used.}

\item{data}{the data to be used.}
}
\value{
The parameter values for running the control chart \code{chart}.
}
\description{
Estimates the parameters used to run a control chart from given
data.
}
\section{Methods (by class)}{
\itemize{
\item \code{SPCchart}: Standard method which simply first applies
PofData to get a model and then xiofP to get the parameters.
}}

