\name{xiofP-methods}
\docType{methods}
\alias{xiofP}
\alias{xiofP-methods}
\alias{xiofP,SPCCUSUMNonparCenterScale-method}
\alias{xiofP,SPCCUSUMNormal-method}
\alias{xiofP,SPCShewNonparCenterScale-method}
\alias{xiofP,SPCShewNormalCenterScale-method}
\title{Methods for Function \code{xiofP} in Package \pkg{spcadjust}}
\description{
Methods for function \code{xiofP} in package \pkg{spcadjust} 
}
\section{Methods}{
\describe{

\item{\code{signature(chart = "SPCCUSUMNonparCenterScale")}}{
}

\item{\code{signature(chart = "SPCCUSUMNormal")}}{
}

\item{\code{signature(chart = "SPCShewNonparCenterScale")}}{
}

\item{\code{signature(chart = "SPCShewNormalCenterScale")}}{
}
}}
\keyword{methods}
